requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
arrow==1.3.0
asyncpg==0.30.0
authx==1.4.1
binaryornot==0.4.4
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
ecdsa==0.19.0
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
text-unidecode==1.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
